# swagger_server/controllers/vens_controller.py
import connexion
from datetime import datetime
from http import HTTPStatus
import logging

from swagger_server.models.problem import Problem  # noqa: E501
from swagger_server.models.resource import Resource  # noqa: E501
from swagger_server.models.ven import Ven  # noqa: E501
from swagger_server.models.resource_request import ResourceRequest  # noqa: E501
from swagger_server.models.ven_request import VenRequest  # noqa: E501
from swagger_server.controllers.subscriptions_controller import subscription_callback  # noqa: E501
from swagger_server.objStore.storageInterface import objStore
from swagger_server import util

# recall this is just a toy VTN
MAX_RESOURCES = 3
# always incrementing
resource_id = 0

def create_ven(body):  # noqa: E501
    """create ven

    Create a new ven. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: Ven
    """
    logging.info(f"create_ven(): ")

    venBody = None
    if connexion.request.is_json:
        venBody = VenRequest.from_dict(connexion.request.get_json())  # noqa: E501

    # object must have unique name
    vens = objStore.search_all("VEN")
    vensList = [v for v in vens if v.ven_name == venBody.ven_name]
    if len(vensList) > 0:
        return [], HTTPStatus.CONFLICT

    now = datetime.now()
    current_time = now.strftime("%H:%M:%S")

    ven = Ven(
        created_date_time=current_time,
        modification_date_time=None,
        object_type='VEN',
        ven_name=venBody.ven_name,
        attributes=venBody.attributes,
        targets=venBody.targets
    )

    if venBody.resources is not None:
        ven.resources = venBody.resources
    else:
        ven.resources = []

    status = objStore.insert(ven)
    if status != HTTPStatus.CREATED:
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"create_ven(): problem={problem}")
        return problem, status

    subscription_callback("VEN", "CREATE", ven)
    return ven, HTTPStatus.CREATED

def delete_ven(ven_id):  # noqa: E501
    """delete  ven

    Delete the ven specified by venID specified in path. # noqa: E501

    :param ven_id: object ID of ven.
    :type ven_id: dict | bytes

    :rtype: Ven
    """

    logging.info(f"delete_ven(): ven_id={ven_id}")
    ven = objStore.remove("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"delete_ven(): problem={problem}")
        return problem, status

    subscription_callback("VEN", "DELETE", ven)

    return ven, HTTPStatus.OK


def search_ven_by_id(ven_id):  # noqa: E501
    """search vens by ID

    Return the ven specified by venID specified in path. # noqa: E501

    :param ven_id: object ID of ven.
    :type ven_id: dict | bytes

    :rtype: Ven
    """
    logging.info(f"search_ven_by_id(): ven_id={ven_id}")
    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"search_ven_by_id(): problem={problem}")
        return problem, status

    subscription_callback("VEN", "READ", ven)

    return ven, HTTPStatus.OK

def search_vens(ven_name=None, target_type=None, target_values=None, skip=None, limit=None):  # noqa: E501
    """search vens

    List all vens. Use skip and pagination query params to limit response size.  # noqa: E501

    :param ven_name: return vens that match requested ven_name
    :type ven_name: list | bytes
    :param targets: return vens that match requested targets
    :type targets: list | bytes
    :param skip: number of records to skip for pagination.
    :type skip: int
    :param limit: maximum number of records to return.
    :type limit: int

    :rtype: List[Ven]
    """
    logging.info(
        f"search_vens(): ven_name={ven_name} target_type={target_type} target_values={target_values} skip={skip} limit={limit}")

    vens = objStore.search_all("VEN")
    logging.debug(f"search_vens(): vens={vens}")
    if type(vens) is not list:
        status = vens
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"search_vens(): problem={problem}")
        return problem, status

    logging.info(f"search_vens(): vens={vens}")
    venList = vens
    if ven_name != None:
        venList = [ven for ven in vens if ven.ven_name == ven_name]
        if len(venList) == 0:
            return venList, HTTPStatus.OK
    venList = util.getTargets(venList, target_type, target_values)
    if skip != None:
        if len(vens) < skip:
            return [], HTTPStatus.OK
        venList = vens[skip:]
    if limit != None:
        venList = venList[:limit]

    logging.debug(f"search_vens(): venList={venList}")

    subscription_callback("VEN", "READ", venList)

    return venList

def update_ven(ven_id, body=None):  # noqa: E501
    """update  ven

    Update the ven specified by venID specified in path. # noqa: E501

    :param ven_id: object ID of ven.
    :type ven_id: dict | bytes
    :param body: ven item to update.
    :type body: dict | bytes

    :rtype: Ven
    """
    logging.info(f"update_ven(): ven_id={ven_id}")

    venBody = None
    if connexion.request.is_json:
        venBody = VenRequest.from_dict(connexion.request.get_json())  # noqa: E501

    if venBody is None:
        problem = Problem(title="Bad Request: No request body", status="400")
        logging.warning(f"update_ven(): problem={problem}")
        return problem, HTTPStatus.BAD_REQUEST

    ven, status = search_ven_by_id(ven_id)
    if ven is None or status == HTTPStatus.NOT_FOUND:
        problem = Problem(title="Not Found: program_id not found", status="404")
        logging.warning(f"update_ven(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    # set modification date time
    now = datetime.now()
    current_time = now.strftime("%H:%M:%S")
    ven.modification_date_time = current_time

    if venBody.ven_name is not None:
        ven.ven_name = venBody.ven_name
    if venBody.targets is not None:
        ven.target = venBody.targets
    if venBody.resources is not None:
        ven.resources = venBody.resources
    if venBody.attributes is not None:
        ven.attributes = venBody.attributes,

    ven = objStore.update("VEN", ven)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"update_ven(): problem={problem}")
        return problem, status

    subscription_callback("VEN", "UPDATE", ven)
    return ven

def create_resource(body, ven_id):  # noqa: E501
    """create resource

    Create a new resource. # noqa: E501

    :param body:
    :type body: dict | bytes
    :param ven_id: Numeric ID of ven.
    :type ven_id: dict | bytes

    :rtype: Resource
     """
    logging.info(f"create_resource(): ven_id={ven_id}")
    global resource_id

    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"create_resource(): problem={problem}")
        return problem, status

    resourceBody = None
    if connexion.request.is_json:
        resourceBody = ResourceRequest.from_dict(connexion.request.get_json())  # noqa: E501

    logging.debug(f"search_all_ven_resources(): ven.resources={ven.resources}")

    # object must have unique name
    resources = ven.resources
    resourceList = [r for r in resources if r.resource_name == resourceBody.resource_name]
    if len(resourceList) > 0:
        return [], HTTPStatus.CONFLICT

    now = datetime.now()
    current_time = now.strftime("%H:%M:%S")

    venResource = Resource(
        id=str(resource_id),
        created_date_time=current_time,
        modification_date_time=None,
        object_type='RESOURCE',
        resource_name=resourceBody.resource_name,
        ven_id = ven_id,
        attributes=resourceBody.attributes,
        targets=resourceBody.targets
    )

    resource_id += 1

    ven.resources.append(venResource)
    ven = objStore.update("VEN", ven)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"update_ven(): problem={problem}")
        return problem, status

    logging.debug(f"create_resource(): venResource={venResource}")

    subscription_callback("RESOURCE", "CREATE", venResource)
    return venResource, HTTPStatus.CREATED

def delete_ven_resource(ven_id, resource_id):  # noqa: E501
    """delete  ven resource

    Delete the ven resource specified by venID and resourceID specified in path. # noqa: E501

    :param ven_id: object ID of the associated ven.
    :type ven_id: dict | bytes
    :param resource_id: object ID of the resource.
    :type resource_id: dict | bytes

    :rtype: Resource
    """
    logging.info(f"delete_ven_resource(): ven_id={ven_id} resource_id={resource_id}")

    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"search_ven_by_id(): problem={problem}")
        return problem, status

    logging.debug(f"delete_ven_resource(): ven={ven}")

    if len(ven.resources) == 0:
        problem = Problem(title="Not Found: no resources in system", status="404")
        logging.warning(f"delete_ven_resource(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    venResource = next((venResource for venResource in ven.resources if venResource.id == resource_id), None)
    if venResource is None:
        problem = Problem(title="Not Found: resource_id not found", status="404")
        logging.warning(f"delete_ven_resource: problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    logging.debug(f"delete_ven_resource(): venResource={venResource}")
    logging.debug(f"delete_ven_resource(): ven.resources={ven.resources}")

    ven.resources.remove(venResource)
    subscription_callback("RESOURCE", "DELETE", venResource)
    return venResource

def search_ven_resources(ven_id, resource_name: str =None, target_type=None, target_values=None, skip=None, limit=None):  # noqa: E501
    """search ven resources

    Return the ven resources specified by venID specified in path. # noqa: E501

    :param ven_id: Numeric ID of ven.
    :type ven_id: dict | bytes
     param resource_name: return resources that match requested resource_name
    :type resource_name: list | bytes
    :param targets: return resources that match requested targets
    :type targets: list | bytes
    :param skip: number of records to skip for pagination.
    :type skip: int
    :param limit: maximum number of records to return.
    :type limit: int

    :rtype: Resource
    """
    logging.info(
        f"search_ven_resources(): resource_name={resource_name} target_type={target_type} target_values={target_values} skip={skip} limit={limit}")
    # TBD: somehow string is received as '[/'name/']' ???
    if resource_name != None and resource_name[0] == '[':
        resource_name = resource_name[2: len(resource_name) - 2]

    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"search_ven_resources(): problem={problem}")
        return problem, status

    resourceList = ven.resources
    logging.info(f"search_ven_resources(): resourceList={resourceList}")
    if resource_name != None:
        # for resource in resourceList:
        #     logging.info(f"resource.resource_name={resource.resource_name} resource_name={resource_name}")

        resourceList = [resource for resource in ven.resources if resource.resource_name == resource_name]
        # logging.info(f"search_ven_resources(): resourceList={resourceList}")
        if len(resourceList) == 0:
            return resourceList, HTTPStatus.OK
    logging.info(f"search_ven_resources(): resourceList={resourceList}")
    resourceList = util.getTargets(resourceList, target_type, target_values)
    if skip != None:
        if len(resourceList) < skip:
            return [], HTTPStatus.OK
        resourceList = resourceList[skip:]
    if limit != None:
        resourceList = resourceList[:limit]
    logging.debug(f"search_ven_resources(): resourceList={resourceList}")

    subscription_callback("RESOURCE", "READ", resourceList)

    return resourceList

def search_ven_resource_by_id(ven_id, resource_id):  # noqa: E501
    """search ven resources by ID

    Return the ven resource specified by venID and resourceID specified in path. # noqa: E501

    :param ven_id: object ID of the associated ven.
    :type ven_id: dict | bytes
    :param resource_id: object ID of the resource.
    :type resource_id: dict | bytes

    :rtype: Resource
    """
    # find index in vens of ven_id
    logging.info(f"search_ven_resource_by_id(): ven_id={ven_id} resource_id={resource_id}")

    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"search_ven_by_id(): problem={problem}")
        return problem, status

    logging.debug(f"search_ven_resource_by_id(): ven={ven}")

    if len(ven.resources) == 0:
        problem = Problem(title="Not Found: no resources in system", status="404")
        logging.warning(f"search_ven_resource_by_id(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    resource = next((resource for resource in ven.resources if resource.id == resource_id), None)
    if resource is None:
        problem = Problem(title="Not Found: resource_id not found", status="404")
        logging.warning(f"search_ven_resource_by_id(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    logging.debug(f"search_ven_resource_by_id(): resource={resource}")

    subscription_callback("RESOURCE", "READ", resource)

    return resource

def update_ven_resource(ven_id, resource_id, body=None):  # noqa: E501
    """update  ven resource

    Update the ven resource specified by venID and resourceID specified in path. # noqa: E501

    :param ven_id: object ID of the associated ven.
    :type ven_id: dict | bytes
    :param resource_id: object ID of the resource.
    :type resource_id: dict | bytes
    :param body: resource item to update.
    :type body: dict | bytes

    :rtype: Resource
    """
    logging.info(f"update_ven_resource(): ven_id={ven_id} resource_id={resource_id}")
    ven = objStore.search("VEN", ven_id)
    if type(ven) is not Ven:
        status = ven
        problem = Problem(title="object Storage issue", status=str(status))
        logging.warning(f"update_ven_resource(): problem={problem}")
        return problem, status

    logging.debug(f"update_ven_resource(): ven={ven}")

    if len(ven.resources) == 0:
        problem = Problem(title="Not Found: no resources in system", status="404")
        logging.warning(f"update_ven_resource(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    resourceBody = None
    if connexion.request.is_json:
        resourceBody = ResourceRequest.from_dict(connexion.request.get_json())  # noqa: E501

    if resourceBody is None:
        problem = Problem(title="Bad Request: No request body", status="400")
        logging.warning(f"update_ven_resource(): problem={problem}")
        return problem, HTTPStatus.BAD_REQUEST

    venResource = next((venResource for venResource in ven.resources if venResource.id == resource_id), None)
    if venResource is None:
        problem = Problem(title="Not Found: resource_id not found", status="404")
        logging.warning(f"update_ven_resource(): problem={problem}")
        return problem, HTTPStatus.NOT_FOUND

    ven.resources.remove(venResource)

    # set modification date time
    now = datetime.now()
    current_time = now.strftime("%H:%M:%S")
    venResource.modification_date_time = current_time

    if resourceBody.resource_name is not None:
        venResource.resource_name = resourceBody.resource_name
    if resourceBody.targets is not None:
        venResource.targets = resourceBody.targets
    if resourceBody.attributes is not None:
        venResource.attributes = resourceBody.attributes,

    logging.debug(f"update_ven_resource(): venResource={venResource}")
    ven.resources.append(venResource)
    subscription_callback("RESOURCE", "UPDATE", venResource)
    return (venResource)